Day 2
Revisited the Dog class I’ve made the previous day, which is just a code representation of a dog which takes a few parameters (like age, name, etc.) and used a while loop to go through an array I’ve created to represent strings of items the dog fetched. Other than that, went through if-else statements, the different loops,
Also created a new project making use of the scanner class allowing the user to enter an age and printing a String depending on what age was input.